// contracts/generated/typescript/models/AITrustProfile.ts
export interface AITrustProfile {
    /**
     *
     * @type {string}
     * @memberof AITrustProfile
     */
    wikiId: string;
    /**
     *
     * @type {string}
     * @memberof AITrustProfile
     */
    actorId: string;
    /**
     *
     * @type {AITrustProfileActorTypeEnum}
     * @memberof AITrustProfile
     */
    actorType: AITrustProfileActorTypeEnum;
    /**
     *
     * @type {string}
     * @memberof AITrustProfile
     */
    actorDisplayName: string;
    /**
     *
     * @type {AITrustProfileTrustLevelEnum}
     * @memberof AITrustProfile
     */
    trustLevel: AITrustProfileTrustLevelEnum;
    /**
     *
     * @type {number}
     * @memberof AITrustProfile
     */
    requiredSamplePercent: number;
    /**
     *
     * @type {boolean}
     * @memberof AITrustProfile
     */
    configured: boolean;
    /**
     *
     * @type {string}
     * @memberof AITrustProfile
     */
    updatedBy?: string;
    /**
     *
     * @type {Date}
     * @memberof AITrustProfile
     */
    createdAt?: Date;
    /**
     *
     * @type {Date}
     * @memberof AITrustProfile
     */
    updatedAt?: Date;
}


/**
 * @export
 */
export const AITrustProfileActorTypeEnum = {
    Ai: 'ai',
    Import: 'import'
} as const;
export type AITrustProfileActorTypeEnum = typeof AITrustProfileActorTypeEnum[keyof typeof AITrustProfileActorTypeEnum];

/**
 * @export
 */
export const AITrustProfileTrustLevelEnum = {
    Untrusted: 'untrusted',
    Assisted: 'assisted',
    Trusted: 'trusted'
} as const;
export type AITrustProfileTrustLevelEnum = typeof AITrustProfileTrustLevelEnum[keyof typeof AITrustProfileTrustLevelEnum];